requirements.txt
# requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
APScheduler==3.10.4
python-dotenv==1.0.0
Django==5.0
psycopg2-binary==2.9.9
Pillow==10.2.0
aiofiles==23.2.1